% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{check_api}
\alias{check_api}
\title{Check if the ScPCA API is reachable}
\usage{
check_api()
}
\value{
TRUE if the API is reachable, otherwise an error is raised.
}
\description{
This function performs a simple GET request to the ScPCA API to verify that it is reachable.
}
\keyword{internal}
