% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computed-files.R
\name{computed_files_filter}
\alias{computed_files_filter}
\title{Helper function to create a filtering list for computed files by format}
\usage{
computed_files_filter(
  format_str = c("SINGLE_CELL_EXPERIMENT", "ANN_DATA", "SPATIAL")
)
}
\arguments{
\item{format_str}{a string indicating the desired format}
}
\value{
a list suitable for passing as the \code{filters} argument to \code{get_computed_file_ids()}.
}
\description{
Helper function to create a filtering list for computed files by format
}
