% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_auth}
\alias{get_auth}
\alias{view_terms}
\title{Get an authorization token from the ScPCA API}
\usage{
get_auth(email, agree = FALSE)

view_terms()
}
\arguments{
\item{email}{The user's email address}

\item{agree}{A logical indicating whether the user agrees to the terms of service}
}
\value{
A string containing the authorization token
}
\description{
\code{get_auth()} allows obtaining an authorization token string from the ScPCA API,
after providing an email address and agreeing to the terms of use.
}
\details{
To view the terms of use before agreeing to them, use \code{view_terms()}, which opens the terms of use page in a web browser.
}
\examples{
\dontrun{
# Get a token (make sure to agree to the terms of service)
auth_token <- get_auth("your.email@example.com", agree = TRUE)
}

\dontrun{
view_terms()
}
}
