% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_ccdl_dataset_detail}
\alias{get_ccdl_dataset_detail}
\title{Fetch a single CCDL dataset by ID}
\usage{
get_ccdl_dataset_detail(id, auth_token)
}
\arguments{
\item{id}{CCDL dataset ID}

\item{auth_token}{API authentication token}
}
\value{
the dataset detail as a list
}
\description{
Fetch a single CCDL dataset by ID
}
\keyword{internal}
