% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computed-files.R
\name{get_computed_file_ids}
\alias{get_computed_file_ids}
\title{Get computed file ids from a sample info list, optionally filtered by criteria}
\usage{
get_computed_file_ids(info_list, filters = list())
}
\arguments{
\item{info_list}{A list object that includes a "computed_files" element,
such as returned by get_sample_info() or get_project_info()}

\item{filters}{A named list of filtering criteria, where names are fields in
the computed_files objects, and values are the desired values to match.
Values can be negated by prefixing with "!". For example, to get all non-spatial
computed files in SingleCellExperiment format, use:
list(format = "SINGLE_CELL_EXPERIMENT", modality = "!SPATIAL").}
}
\value{
a character vector of computed file ids matching the filtering criteria
}
\description{
Get computed file ids from a sample info list, optionally filtered by criteria
}
\keyword{internal}
