% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_project_libraries}
\alias{get_project_libraries}
\title{Get metadata for all libraries in a given project}
\usage{
get_project_libraries(project_id, auth_token)
}
\arguments{
\item{project_id}{The ScPCA project ID (e.g. "SCPCP000001")}

\item{auth_token}{An authorization token obtained from \code{get_auth()}}
}
\value{
A data frame (tibble) of library metadata for the specified project.
}
\description{
This function downloads and reads the library metadata file for a given ScPCA project.
The data frame returned will be the same as the project metadata file available from the
ScPCA Portal website for each project, including information about each library
that is part of the project.
}
\examples{
\dontrun{
# First get an auth token
token <- get_auth("me@email.net", agree = TRUE)
# Get library metadata for a specific project
libraries_df <- get_project_libraries("SCPCP000001", token)
}

}
