% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_project_metadata_url}
\alias{get_project_metadata_url}
\title{Get project metadata download URL}
\usage{
get_project_metadata_url(project_id, auth_token)
}
\arguments{
\item{project_id}{The ScPCA project ID (e.g. "SCPCP000001")}

\item{auth_token}{An authorization token obtained from \code{get_auth()}}
}
\value{
A signed download URL for the project metadata file as would be found
from the ScPCA Portal.
}
\description{
Get project metadata download URL
}
\examples{
\dontrun{
# First get an auth token
token <- get_auth("me@email.net", agree = TRUE)
# Get metadata for a specific project
project_info <- get_project_metadata_url("SCPCP000001", token)
}
}
\keyword{internal}
