% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_project_samples}
\alias{get_project_samples}
\title{Get a data frame of all samples in a given project}
\usage{
get_project_samples(project_id, simplify = TRUE)
}
\arguments{
\item{project_id}{The project ID (e.g. "SCPCP000001")}

\item{simplify}{A logical indicating whether to simplify the resulting data frame
by removing list columns. Default is TRUE.}
}
\value{
a data frame of sample information for the specified project from the ScPCA API.
}
\description{
This function retrievs a data frame of all biological samples associated with a SCPCA project,
including sample-level metadata. By default, list columns are removed to create
a simplified data frame, but this can be disabled by setting \code{simplify = FALSE}.
The unsimplified data frame contains nested list columns with additional details,
such as the experimental modalities associated with each sample.
}
\examples{
\dontrun{
# Get sample info for a specific project
samples_df <- get_project_samples("SCPCP000001")

# Get sample info without simplifying
samples_df_full <- get_project_samples("SCPCP000001", simplify = FALSE)
}
}
