% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{iterate_scpca}
\alias{iterate_scpca}
\title{Helper function for iterating through paginated ScPCA API results}
\usage{
iterate_scpca(resp, req)
}
\arguments{
\item{resp}{httr2 response object}

\item{req}{httr2 request object}
}
\value{
updated httr2 request object for the next page, or NULL if there are no more pages
}
\description{
Helper function for iterating through paginated ScPCA API results
}
