% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/downloads.R
\name{normalize_format}
\alias{normalize_format}
\title{Internal helper to validate and normalize formats for the ScPCA API}
\usage{
normalize_format(format)

normalize_format(format)
}
\arguments{
\item{format}{The input format string}
}
\value{
The normalized format string for API use

The normalized format string for API use
}
\description{
Note that while spatial format strings are included here, the API does not
currently accommodate a format code for spatial data; that is designated in the
modality field for individual samples or precomputed datasets.
}
\keyword{internal}
