% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{parse_download_file}
\alias{parse_download_file}
\title{Get the base filename from a ScPCA portal download URL}
\usage{
parse_download_file(scpca_url)
}
\arguments{
\item{scpca_url}{The ScPCA portal download URL}
}
\value{
the download filename
}
\description{
(this may become obsolete if we get download filenames in the API response)
}
\keyword{internal}
