% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{scpca_projects}
\alias{scpca_projects}
\title{Get data frame of all ScPCA projects}
\usage{
scpca_projects()
}
\value{
a data frame of project information from the ScPCA API.
}
\description{
Get data frame of all ScPCA projects
}
