% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{scpca_request}
\alias{scpca_request}
\title{Base request object for ScPCA API}
\usage{
scpca_request(resource, body = list(), auth_token = "", ...)
}
\arguments{
\item{resource}{API resource to query, e.g. "projects"}

\item{body}{optional named list to include as JSON body in the request}

\item{auth_token}{optional API authentication token}

\item{...}{additional query parameters to include in the request}
}
\value{
a httr2 request object
}
\description{
Base request object for ScPCA API
}
