% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{view_terms}
\alias{view_terms}
\title{Open the ScPCA Terms of Use page in a browser}
\usage{
view_terms()
}
\value{
Invisibly returns the URL that was opened (character scalar).
}
\description{
This opens the ScPCA terms of use URL in the user's default web browser.
}
\examples{
\dontrun{
view_terms()
}
}
