% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{download_project}
\alias{download_project}
\title{Download a project's data files from the ScPCA Portal}
\usage{
download_project(
  project_id,
  auth_token,
  destination = "scpca_data",
  format = "sce",
  merged = FALSE,
  include_multiplexed = NULL,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{project_id}{The ScPCA project ID (e.g. "SCPCP000001")}

\item{auth_token}{An authorization token obtained from \code{get_auth()}}

\item{destination}{The path to the directory where the unzipped file directory should be saved.
Default is "scpca_data".}

\item{format}{The desired file format, either "sce" (SingleCellExperiment),
"anndata" (AnnData/H5AD), or "spatial" (for spatial data in Space Ranger format).
Default is "sce".}

\item{merged}{Download merged data files, if available.
Default is FALSE.}

\item{include_multiplexed}{Include multiplexed samples, if available.
Default is TRUE for SingleCellExperiment and FALSE for AnnData and spatial samples,
where multiplexed data are not available.}

\item{overwrite}{Whether to overwrite existing directories if they already exist. Default is FALSE.}

\item{quiet}{Whether to suppress download progress messages. Default is FALSE.}
}
\value{
a vector of file paths for the downloaded files (invisibly)
}
\description{
Download a project's data files from the ScPCA Portal
}
\examples{

\dontrun{
# Get a token first
auth_token <- get_auth("me@email.net", agree = TRUE)
# Then ask for a sample download
download_project("SCPCS000001", auth_token, destination = "scpca_data", format = "sce")

# Downloading merged files in AnnData format
download_project(
  "SCPCS000001",
  auth_token,
  destination = "scpca_data",
  format = "anndata",
  merged = TRUE
)
}
}
