% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_auth}
\alias{get_auth}
\title{Get an authorization token from the ScPCA API}
\usage{
get_auth(email, agree = FALSE)
}
\arguments{
\item{email}{The user's email address}

\item{agree}{A logical indicating whether the user agrees to the terms of service}
}
\value{
A string containing the authorization token
}
\description{
Get an authorization token from the ScPCA API
}
\examples{
\dontrun{
# Get a token (make sure to agree to the terms of service)
auth_token <- get_auth("me@email.net", agree = TRUE)
}

}
