% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{get_project_info}
\alias{get_project_info}
\title{Get project metadata by project ID}
\usage{
get_project_info(project_id, simplifyVector = FALSE)
}
\arguments{
\item{project_id}{the project ID (e.g. "SCPCP000001")}

\item{simplifyVector}{passed to resp_body_json to simplify the returned JSON structure}
}
\value{
a list of project metadata from the ScPCA API.
}
\description{
Get project metadata by project ID
}
