% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{scpca_projects}
\alias{scpca_projects}
\title{Get data frame of all ScPCA projects}
\usage{
scpca_projects(simplify = TRUE)
}
\arguments{
\item{simplify}{A logical indicating whether to simplify the resulting data frame by removing list columns. Default is TRUE.}
}
\value{
a data frame of project information from the ScPCA API.
}
\description{
Get data frame of all ScPCA projects
}
