% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{scpca_request}
\alias{scpca_request}
\title{Base request object for ScPCA API}
\usage{
scpca_request(resource, limit = 20, offset = 0)
}
\arguments{
\item{resource}{API resource to query, e.g. "projects"}

\item{limit}{number of records to return (default 20)}

\item{offset}{number of records to skip (default 0)}
}
\value{
a httr2 request object
}
\description{
Base request object for ScPCA API
}
